refactor(web): extract backtesting mutations
