+++ web/frontend/src/app/backtesting/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useQuery, useMutation } from "@tanstack/react-query";
+import { useQuery } from "@tanstack/react-query";
 import { toast } from "sonner";
 import { ToolLayout } from "@/components/common/tool-layout";
 import { EmptyState } from "@/components/common/empty-state";
@@ -12,19 +12,15 @@ import {
 import {
     Activity,
 } from "lucide-react";
-import { apiGet, apiPost } from "@/lib/api";
+import { apiGet } from "@/lib/api";
 import { formatCurrencyPrecise } from "@/lib/format";
 import { useBacktestStore } from "@/stores/backtest-store";
 import type {
     StrategyInfo,
     BacktestRequest,
-    BacktestResponse,
-    SaveExperimentRequest,
-    SaveExperimentResponse,
 } from "@/types/api";
 import type {
     ComparisonPortfolio,
-    ComparisonRun,
 } from "@/stores/backtest-store";
 import {
     PORTFOLIO_PRESETS,
@@ -42,6 +38,7 @@ import {
     buildComparisonBacktestRequest,
 } from "@/app/backtesting/backtesting-request-builder";
 import { useBacktestingFormActions } from "@/app/backtesting/use-backtesting-form-actions";
+import { useBacktestingMutations } from "@/app/backtesting/use-backtesting-mutations";
 import { useBacktestingPortfolioActions } from "@/app/backtesting/use-backtesting-portfolio-actions";
 import { useBacktestingRunActions } from "@/app/backtesting/use-backtesting-run-actions";
 
@@ -119,80 +116,6 @@ export default function BacktestingPage() {
             .includes(query);
     });
 
-    const mutation = useMutation({
-        mutationFn: (req: BacktestRequest) =>
-            apiPost<BacktestResponse>("/api/backtesting/run", req, 120000),
-        onSuccess: (_data, variables) => {
-            setLastRunRequest(variables);
-            setSavedExperiment(null);
-            toast.success("Backtest complete");
-        },
-        onError: (e) => toast.error(`Backtest failed: ${e.message}`),
-    });
-
-    const saveExperimentMutation = useMutation({
-        mutationFn: (req: SaveExperimentRequest) =>
-            apiPost<SaveExperimentResponse>("/api/experiments/save", req),
-        onSuccess: (data) => {
-            setSavedExperiment(data);
-            toast.success("Experiment saved");
-        },
-        onError: (e) => toast.error(`Save failed: ${e.message}`),
-    });
-
-    const comparisonMutation = useMutation({
-        mutationFn: async (portfolios: ComparisonPortfolio[]) => {
-            const baseRequest = buildBacktestRequest();
-            if (!baseRequest) {
-                throw new Error("Fix the current run configuration before comparing portfolios.");
-            }
-
-            const runs = await Promise.all(
-                portfolios.map(async (portfolio): Promise<ComparisonRun> => {
-                    const request = buildComparisonRequest(portfolio, baseRequest);
-                    try {
-                        const comparisonResult = await apiPost<BacktestResponse>(
-                            "/api/backtesting/run",
-                            request,
-                            120000,
-                        );
-                        return {
-                            portfolio,
-                            request,
-                            result: comparisonResult,
-                        };
-                    } catch (error) {
-                        return {
-                            portfolio,
-                            request,
-                            result: null,
-                            error:
-                                error instanceof Error
-                                    ? error.message
-                                    : "Comparison run failed",
-                        };
-                    }
-                }),
-            );
-            return runs;
-        },
-        onSuccess: (runs) => {
-            setComparisonRuns(runs);
-            const firstRequest = runs.find((run) => run.request)?.request;
-            if (firstRequest) {
-                setLastComparisonRequest(firstRequest);
-            }
-            const failed = runs.filter((run) => run.error).length;
-            if (failed > 0) {
-                toast.error(`${failed} comparison run${failed === 1 ? "" : "s"} failed`);
-            } else {
-                toast.success("Comparison complete");
-            }
-            setActiveResultTab("comparison");
-        },
-        onError: (e) => toast.error(`Comparison failed: ${e.message}`),
-    });
-
     const {
         applyStrategy,
         updateCostAssumption,
@@ -222,38 +145,6 @@ export default function BacktestingPage() {
         setMissingDataPolicy,
         setCostAssumptions,
         setParams,
-    });
-
-    const {
-        applyPortfolioPreset,
-        updatePortfolioTicker,
-        updatePortfolioWeight,
-        addPortfolioAsset,
-        removePortfolioAsset,
-        equalizePortfolioWeights,
-        normalizePortfolioWeights,
-        clearPortfolioAssets,
-        handleSavePortfolio,
-        applySavedPortfolio,
-        removeSavedPortfolio,
-        handleAddCurrentPortfolioToComparison,
-        handleAddPresetToComparison,
-        handleAddSavedToComparison,
-        removeComparisonPortfolio,
-        handleRunComparison,
-    } = useBacktestingPortfolioActions({
-        strategy,
-        params,
-        portfolioAssets,
-        isAllocationStrategy,
-        allocationIsBalanced,
-        comparisonPortfolios,
-        setPortfolioAssets,
-        setSavedPortfolios,
-        setComparisonPortfolios,
-        setParams,
-        applyStrategy,
-        onRunComparison: comparisonMutation.mutate,
     });
 
     const buildBacktestRequest = (): BacktestRequest | null => {
@@ -292,6 +183,52 @@ export default function BacktestingPage() {
         baseRequest: BacktestRequest,
     ): BacktestRequest =>
         buildComparisonBacktestRequest({ portfolio, baseRequest, params });
+
+    const {
+        mutation,
+        saveExperimentMutation,
+        comparisonMutation,
+    } = useBacktestingMutations({
+        buildBacktestRequest,
+        buildComparisonRequest,
+        setLastRunRequest,
+        setLastComparisonRequest,
+        setSavedExperiment,
+        setComparisonRuns,
+        setActiveResultTab,
+    });
+
+    const {
+        applyPortfolioPreset,
+        updatePortfolioTicker,
+        updatePortfolioWeight,
+        addPortfolioAsset,
+        removePortfolioAsset,
+        equalizePortfolioWeights,
+        normalizePortfolioWeights,
+        clearPortfolioAssets,
+        handleSavePortfolio,
+        applySavedPortfolio,
+        removeSavedPortfolio,
+        handleAddCurrentPortfolioToComparison,
+        handleAddPresetToComparison,
+        handleAddSavedToComparison,
+        removeComparisonPortfolio,
+        handleRunComparison,
+    } = useBacktestingPortfolioActions({
+        strategy,
+        params,
+        portfolioAssets,
+        isAllocationStrategy,
+        allocationIsBalanced,
+        comparisonPortfolios,
+        setPortfolioAssets,
+        setSavedPortfolios,
+        setComparisonPortfolios,
+        setParams,
+        applyStrategy,
+        onRunComparison: comparisonMutation.mutate,
+    });
 
     const result = mutation.data;
     const {
